LayoutTests/fast/js/out-of-memory.html

<div>This page should throw an out of memory exception.</div>
<script>
if (window.layoutTestController)
  layoutTestController.dumpAsText();

var object = new Array;
var counter = 0;
try {
  while(true) {
    object[counter++] = new Array;
  }
} catch(e) {
  alert("FAILED! caught an Out Of Memory Exception!!  " + e);
}
alert("FAILED!  No out of memory exception was thrown!");
</script>
