LayoutTests/fast/js/out-of-memory-expected.txt

CONSOLE MESSAGE: line 10: Error: Out of memory
This page should throw an out of memory exception.
